requirements.txt
runpod>=1.3.0
torch>=2.0.0
torchvision
torchaudio
requests
Pillow
opencv-python
accelerate
transformers
diffusers
xformers